Pincode Routes and Repo Added

+++ src/data/model/Pincode.kt
@@ -1,7 +1,7 @@
 package com.example.data.model
 
 data class Pincode(
-    val pincode : Long,
+    val pincode : Int,
     val area_name : String,
     val delivery_charge : Int
 )
